Changed tab order and minimized top right corner shortcuts
